Fix pagination bug and remove products table
Replaced the table with custom cards. Looks like the
table API does not support dynamic calls like the
ones being used in the component.

Also added the Flex Layout module to make more responsive
the products list UI.

It is left to add more information to the small cards,
the full details will be shown in another component.