Do not add hidden values in the HTML
Hidden values will be kept on the controller's attributes, instead of
adding them to the HTML as hidden values.

Refactored the ModelEditor class with updated logic and, with help of
the Form component, removed the common form logic in the products
module.

Also added some detected TODOs after the refactoring.